Created the project
This is where the first stage of this project starts. Hopefully, I can
finish the life engine this time. I'm interested in how the life works,
and how you can simulate it virtually through programming, and pretty
get the similar experience out of it. It's going to be a very difficult
process, but I'll keep working on it until it's on a release stage.